requirements.txt
langchain==0.3.23
langchain-openai==0.3.14
python-dotenv==1.0.1
faiss-cpu==1.7.4
tiktoken==0.6.0
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
beautifulsoup4==4.12.3
requests==2.31.0
python-multipart==0.0.9
openai==1.14.0